requirements.txt
# AutoXHS 依赖
# =====================================

# 核心依赖
openai>=1.12.0
python-dotenv>=1.0.1
requests>=2.31.0

# 小红书相关
xhs>=0.2.9
playwright>=1.41.0
qrcode>=7.4.2

# 工具库
pydantic>=2.0.0
